+++ Sources/NSAttributedStringBuilder/Components/Link.swift
@@ -9,15 +9,14 @@ public typealias Link = NSAttributedString.Link
 extension NSAttributedString {
     public struct Link: Component {
         public let string: String
-        public let url: URL
+        public let url: URL?
         public let attributes: Attributes
 
-        public init(_ string: String, url: URL, attributes: Attributes = [:]) {
+        public init(_ string: String, url: URL?, attributes: Attributes = [:]) {
             self.string = string
             self.url = url
-
             var attributes = attributes
-            attributes[.link] = url
+            if let url = url { attributes[.link] = url }
             self.attributes = attributes
         }