+++ Sources/NSAttributedStringBuilder/Components/StaticComponents.swift
@@ -2,7 +2,8 @@ import Foundation
 
 public typealias Empty = NSAttributedString.Empty
 public typealias Space = NSAttributedString.Space
-public typealias LineBreak = NSAttributedString.LineBreak
+public typealias Break = NSAttributedString.Break
+public typealias Tab = NSAttributedString.Tab
 
 extension NSAttributedString {
     public struct Empty: Component {
@@ -17,8 +18,14 @@ extension NSAttributedString {
         public init() { }
     }
 
-    public struct LineBreak: Component {
+    public struct Break: Component {
         public let string: String = "\n"
+        public let attributes: Attributes = [:]
+        public init() { }
+    }
+    
+    public struct Tab: Component {
+        public let string: String = "\t"
         public let attributes: Attributes = [:]
         public init() { }
     }