+++ SwiftUISampleApp/AttributedTextSample/AttributedText.swift
@@ -13,16 +13,16 @@ final public class AttributedText: UIViewRepresentable {
         self.init(builder())
     }
 
-    public func makeUIView(context: UIViewRepresentableContext<AttributedText>) -> UITextView {
-        let textView = UITextView(frame: .zero)
-        textView.attributedText = self.attributedString
-        textView.isEditable = false
-        textView.backgroundColor = .clear
-        textView.textAlignment = .center
-        return textView
+    public func makeUIView(context: UIViewRepresentableContext<AttributedText>) -> UILabel {
+        let label = UILabel(frame: .zero)
+        label.attributedText = self.attributedString
+        label.backgroundColor = .clear
+        label.textAlignment = .natural
+        label.numberOfLines = 0
+        return label
     }
 
-    public func updateUIView(_ textView: UITextView, context: UIViewRepresentableContext<AttributedText>) {
+    public func updateUIView(_ textView: UILabel, context: UIViewRepresentableContext<AttributedText>) {
         textView.attributedText = self.attributedString
     }
 }
@@ -30,11 +30,12 @@ final public class AttributedText: UIViewRepresentable {
 #if DEBUG
 struct AttributedText_Previews : PreviewProvider {
     static var previews: some View {
+        ScrollView {
         AttributedText {
+            Tab()
             ImageAttachment(UIImage(named: "Swift_logo_color_rgb.jpg")!, size: CGSize(width: 90, height: 90))
-            LineBreak()
-                .lineSpacing(20)
-            AText("Hello SwiftUI")
+            Break()
+            Content("Hello SwiftUI")
                 .backgroundColor(.red)
                 .baselineOffset(10)
                 .font(.systemFont(ofSize: 20))
@@ -47,10 +48,22 @@ struct AttributedText_Previews : PreviewProvider {
                 .strikethrough(style: .patternDash, color: .black)
                 .stroke(width: -2, color: .green)
                 .underline(.patternDashDotDot, color: .cyan)
-            LineBreak()
-            AText(" with fun")
+                .alignment(.center)
                 .paragraphSpacing(10, before: 60)
-                .alignment(.right)
+            Break()
+            Content("With fun!")
+                .paragraphSpacing(10, before: 60)
+                .alignment(.center)
+            Break()
+            Content("Scroll down...")
+                .paragraphSpacing(10, before: 60)
+                .alignment(.center)
+            Break()
+            Content("End!")
+                .backgroundColor(.red)
+                .paragraphSpacing(10, before: 360)
+                .alignment(.center)
+        }
         }
     }
 }